Move the optimizer utils to common graph util folder
