+++ src/graph/optimizer/rule/IndexFullScanBaseRule.cpp
@@ -9,9 +9,9 @@
 #include "graph/optimizer/OptContext.h"
 #include "graph/optimizer/OptGroup.h"
 #include "graph/optimizer/OptRule.h"
-#include "graph/optimizer/OptimizerUtils.h"
 #include "graph/planner/plan/Query.h"
 #include "graph/planner/plan/Scan.h"
+#include "graph/util/OptimizerUtils.h"
 #include "interface/gen-cpp2/storage_types.h"
 
 using nebula::graph::IndexScan;
