+++ src/graph/optimizer/rule/OptimizeTagIndexScanByFilterRule.cpp
@@ -8,11 +8,11 @@
 #include "graph/context/QueryContext.h"
 #include "graph/optimizer/OptContext.h"
 #include "graph/optimizer/OptGroup.h"
-#include "graph/optimizer/OptimizerUtils.h"
 #include "graph/optimizer/rule/IndexScanRule.h"
 #include "graph/planner/plan/PlanNode.h"
 #include "graph/planner/plan/Scan.h"
 #include "graph/util/ExpressionUtils.h"
+#include "graph/util/OptimizerUtils.h"
 
 using nebula::graph::Filter;
 using nebula::graph::OptimizerUtils;
